Toolkit on how to make Is It Normal
This toolkit is an example for my 460/560 Serious STEM Games course. I've asked graduate students to create toolkits that describe their key design decisions and explain some of the steps it took to make their final projects, and this does that for the Is It Normal game which you can download here as well.